Only remove the original path if it has changed.
When uploading images, we move the raw image to a semi-predictable
path based on the user's team, but using the original file extension
This image is then read, converted & re-output to a fixed path.
As these two paths could be the same, such that when the 'original'
was removed, we in fact removed the final image. Avoid this by checking
the paths are different before removing them, and add a test that there
is an image uploaded, using images for both possible cases.

+++ modules/team.php
@@ -134,8 +134,11 @@ class TeamModule extends Module
 		$dest = path_change_extension($path, 'png');
 		$image->resizeInto($width, $height, $dest);
 
-		// remove the original
-		unlink($path);
+		// remove the original, if different
+		if ($path != $dest)
+		{
+			unlink($path);
+		}
 
 		// update the status store
 		$status = new TeamStatus($team);
